After implementing the user controller

+++ src/main/java/com/Zorvyn/finance_data_service/dto/response/DashboardResponse.java
@@ -3,15 +3,16 @@ package com.Zorvyn.finance_data_service.dto.response;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 @Getter
 @Setter
 public class DashboardResponse {
-    private Double totalIncome;
-    private Double totalExpense;
-    private Double netBalance;
+    private BigDecimal totalIncome;
+    private BigDecimal totalExpense;
+    private BigDecimal netBalance;
     private Map<String, Double> categoryWiseTotals;
     private List<RecordResponse> recentActivity;
 }
